requirements.txt
google-generativeai
langgraph
deep_translator
plotly
gradio
openai
transformers
torch
pandas
matplotlib
